+++ backend/emm-sale/src/main/java/com/emmsale/login/utils/GithubClient.java
@@ -1,4 +1,4 @@
-package com.emmsale.login.application;
+package com.emmsale.login.utils;
 
 import com.emmsale.login.application.dto.GithubAccessTokenRequest;
 import com.emmsale.login.application.dto.GithubAccessTokenResponse;
